requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-Cors==3.0.10
psycopg2-binary==2.9.6
python-dotenv==1.0.0
